% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaod-package.R
\docType{package}
\name{jaod-package}
\alias{jaod-package}
\alias{jaod}
\title{jaod}
\description{
Client for the Directory of Open Access Journals (DOAJ)
\url{https://doaj.org/}.
}
\section{DOAJ Information}{

DOAJ is a (quoting) "community-curated list of open access journals and aims
to be the starting point for all information searches for quality, peer
reviewed open access material."

DOAJ API documentation can be found at \url{https://doaj.org/api/v2/docs}.

DOAJ API FAQ: \url{https://doaj.org/docs/faq/#apikey}
}

\section{Package API}{

\itemize{
\item \code{\link[=jaod_article]{jaod_article()}} - Get an article by ID
\item \code{\link[=jaod_article_search]{jaod_article_search()}} - Search for articles
\item \code{\link[=jaod_journal]{jaod_journal()}} - Get a journal by ID
\item \code{\link[=jaod_journal_search]{jaod_journal_search()}} - Search for journals
}
}

\author{
Scott Chamberlain
}
\keyword{package}
