% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaod_article_search.R
\name{jaod_article_search}
\alias{jaod_article_search}
\title{Search for articles}
\usage{
jaod_article_search(query, page = 1, pageSize = 10, sort = NULL, ...)
}
\arguments{
\item{query}{(character) query terms}

\item{page}{(integer) Which page of the results you wish to see. Default: 1}

\item{pageSize}{(integer) How many results per page you wish to see.
Default: 10}

\item{sort}{(character) one of asc or desc, or sort by a specific field,
either asc or desc, see examples}

\item{...}{curl options passed on to \code{\link[crul]{HttpClient}},
see `?curl::curl_options` for help on curl options}
}
\value{
a tibble (data.frame)
}
\description{
Search for articles
}
\examples{
\dontrun{
out <- jaod_article_search(query = "license:CC-BY")
out
out$results

## get DOIs or other article IDs
names(out$results$bibjson.identifier) <- out$results$id
dplyr::bind_rows(out$results$bibjson.identifier, .id = "doaj_id")

# more egs
jaod_article_search(query = "issn:1544-9173")
jaod_article_search(query = "publisher:dove")
}
}
