% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaod_journal_search.R
\name{jaod_journal_search}
\alias{jaod_journal_search}
\title{Search for journals}
\usage{
jaod_journal_search(query, page = 1, pageSize = 10, sort = NULL, ...)
}
\arguments{
\item{query}{(character) query terms}

\item{page}{(integer) Which page of the results you wish to see. Default: 1}

\item{pageSize}{(integer) How many results per page you wish to see.
Default: 10}

\item{sort}{(character) one of asc or desc, or sort by a specific field,
either asc or desc, see examples}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\value{
a tibble (data.frame)
}
\description{
Search for journals
}
\examples{
\dontrun{
out <- jaod_journal_search(query = "bibjson.keywords:heritage")
out
out$results

jaod_journal_search(query = "issn:1544-9173")
jaod_journal_search(query = "publisher:dove")
}
}

